исправлены ошибки.

+++ src/main/java/team/skyprojava/websitebackend/controller/UserController.java
@@ -175,7 +175,7 @@ public class UserController {
     }
 
     @PutMapping("{id}/updateRole")
-    public ResponseEntity<UserDto> updateRole(@PathVariable("id") long id, Role role) {
+    public ResponseEntity<UserDto> updateRole(@PathVariable("id") int id, Role role) {
         logger.info("Request for update user role");
         return ResponseEntity.ok(userService.updateRole(id, role));
     }
